requirements.txt
fivetran
apache-airflow
pandas
scikit-learn
snowflake-connector-python
dbt-core
dbt-snowflake
